+++ tests/unit/models/test_evc_base.py
@@ -257,7 +257,7 @@ class TestEVC():  # pylint: disable=too-many-public-methods, no-member
         }
         update_dict = {"queue_id": 3}
         evc = EVC(**attributes)
-        _, redeploy = evc.update(**update_dict)
+        _, redeploy, __ = evc.update(**update_dict)
         assert redeploy
 
     @patch("napps.kytos.mef_eline.models.EVC.sync")
@@ -273,7 +273,7 @@ class TestEVC():  # pylint: disable=too-many-public-methods, no-member
         }
         update_dict = {"queue_id": None}
         evc = EVC(**attributes)
-        _, redeploy = evc.update(**update_dict)
+        _, redeploy, __ = evc.update(**update_dict)
         assert redeploy
 
     def test_update_different_tag_lists(self):
@@ -769,3 +769,51 @@ class TestEVC():  # pylint: disable=too-many-public-methods, no-member
         with pytest.raises(ValueError):
             evc.update(**{"primary_path": new_primary_path})
         mock_use_tags.assert_not_called()
+
+    @patch("napps.kytos.mef_eline.models.EVC._get_unis_use_tags")
+    @patch("napps.kytos.mef_eline.models.EVC.sync")
+    def test_from_intra_to_inter(self, _mock_sync, mock_use_tags):
+        """Test _tag_lists_equal"""
+        mock_use_tags.return_value = False, False
+        uni_a = get_uni_mocked(is_valid=True)
+        uni_z = get_uni_mocked(is_valid=True)
+        uni_z.interface.switch = uni_a.interface.switch
+        primary_path = Path([])
+        attributes = {
+            "controller": get_controller_mock(),
+            "name": "circuit_name",
+            "dynamic_backup_path": True,
+            "primary_path": primary_path,
+            "enable": True,
+            "uni_a": uni_a,
+            "uni_z": uni_z,
+        }
+        evc = EVC(**attributes)
+        update_dict = {"uni_z": get_uni_mocked(is_valid=True)}
+        _, __, force_removal = evc.update(**update_dict)
+        assert force_removal is True
+
+    @patch("napps.kytos.mef_eline.models.EVC._get_unis_use_tags")
+    @patch("napps.kytos.mef_eline.models.EVC.sync")
+    def test_from_inter_to_intra(self, _mock_sync, mock_use_tags):
+        """Test _tag_lists_equal"""
+        mock_use_tags.return_value = False, False
+        uni_a = get_uni_mocked(is_valid=True)
+        uni_z = get_uni_mocked(is_valid=True)
+        uni_a_switch = uni_a.interface.switch
+        primary_path = Path([])
+        attributes = {
+            "controller": get_controller_mock(),
+            "name": "circuit_name",
+            "dynamic_backup_path": True,
+            "primary_path": primary_path,
+            "enable": True,
+            "uni_a": uni_a,
+            "uni_z": uni_z,
+        }
+        evc = EVC(**attributes)
+        new_uni_z = get_uni_mocked(is_valid=True)
+        new_uni_z.interface.switch = uni_a_switch
+        update_dict = {"uni_z": new_uni_z}
+        _, __, force_removal = evc.update(**update_dict)
+        assert force_removal is False